pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "minsearch>=0.0.10",
    "pandas>=3.0.1",
    "pydantic-ai>=1.73.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]